cgraph is_anonymous: handle default ID discipline as a special case
The use of `irrelevant_subgraph` in `node_in_subg` shows up as a hot spot when
profiling the ec_pairs-12.gv graph.¹ There was significant unnecessary work
being performed here:
  1. `agnameof` constructs the name of the graph, but we only need to check the
     first character.
  2. `agnameof` anticipates the input object being of arbitrary type, but we
     know it is a graph here.
  3. `agnameof` does some extra processing that is only relevant to edges, yet
     we (again) know the input is a graph.

Taking advantage of this, we can handle the common case in its own path. When
the default ID discipline is in use (almost always), we perform the core
relevant actions that are cheaper than running `agnameof`. Note that we can
skip the internal map lookup (`aginternalmapprint`) because only anonymous
graphs are added to this. If we miss in the main lookup (`id % 2 != 0`) then we
have an anonymous graph, regardless of the internal map lookup.

This is a slight speed up:

  ┌────────────────────┬────────────────────┬────────┐
  │       before       │        after       │  diff  │
  ├────────────────────┼────────────────────┼────────┤
  │  1h03m59s runtime  │   55m34s runtime   │ ~-13%  │
  └────────────────────┴────────────────────┴────────┘

¹ `ccomps ec_pairs-12.gv` using the graph from #2593.

Gitlab: #2593
Reported-by: Diomidis Spinellis