add a test case for 'splines=ortho', 'concentrate=true' crash
Bisect reveals this was fixed in 9f670e8ee818cef72d9e96c6257e0a44fc0a6886. But
since it seems the correlation between that change and this problem was unknown
at the time, it seems wise to introduce a test case to ensure we do not regress
this behavior in future.

Gitlab: #2183
Reported-by: Henry Giddings