pack: lift 'GRID' macro into a function
Contemporary compilers do not need this to be a macro in order to inline it.
Making it a function gives us stronger type safety. Separating the `ceil` call
and its cast also squashes 6 compiler warnings like:

  pack.c: In function ‘genBox’:
  pack.c:34:20: warning: cast from function call of type ‘double’ to non-matching type ‘int’ [-Wbad-function-cast]
     34 | #define GRID(x,s) ((int)ceil((x)/(s)))
        |                    ^
  pack.c:252:9: note: in expansion of macro ‘GRID’
    252 |     W = GRID(bb0.UR.x - bb0.LL.x + 2 * margin, ssize);
        |         ^~~~

The utility of -Wbad-function-cast is debatable, as it mostly warns about
situations like this where the truncation is deliberate.