Debian: fix package dependencies
Many of the Debian packages built in CI had insufficient dependencies. For
example, `graphviz` missing a dependency on `libgraphviz4`. This permitted users
to install the former without the latter. On a clean system, this resulted in
Graphviz binaries that were unrunnable due to missing libraries. However even
more problematically, on a system with a prior installation of `libgraphviz4`
the binaries would be runnable but would misreport their version. When trying to
run functionality that differed between their intended `libgraphviz4` dependent
and the installed one they would likely crash.

Though this was reported on WSL, this affected all Debian-based platforms.

Gitlab: fixes #2466
Reported-by: Edmond Soun