gvpr: fix: remove claim that 'gvpr -c ""' is equivalent to 'nop'
Even in the first revision of Graphviz, this does not appear to have been true.
The logic for whether a program should do anything was near-identical to what
exists today:

  int walksGraph(comp_prog * p)
  {
    return (p->n_nstmts || p->n_estmts);
  }
  …
  int usesGraph(comp_prog * p)
  {
    return (walksGraph(p) || p->begg_stmt || p->endg_stmt);
  }

It is possible to implement a special case in GVPR that makes the man page claim
true. But there seems little point to this when users can simply run `nop`
instead.

Gitlab: fixes #2584
Reported-by: Phormio