pango plugin pango_textlayout: squash -Wfloat-equal warning
The condition here is checking whether the cached font is the same as the
current one being looked up. So we really only want to match when the one being
looked up exactly matches the prior request.

With this change, the Pango plugin builds warning-free with the exception of the
use of deprecated APIs.