gvpr: fix 'depath' to avoid acting on already deleted nodes
This program was traversing the graph and deleting nodes in a single pass. A
side effect of this was that it could find itself acting on a node it had
previously deleted. With a self-loop (`A -> A`), the logic was effectively:

  n = «some node»
  nxt = next(n)
  delete n
  «use nxt» // ← no accounting for nxt aliasing already-deleted n

Even with graphs that did not contain self-loops, problems could still occur.
The implicit loop in this program was effectively:

  n = «some node»
  nxt = next(n)
  while (…) {
    «some stuff, deleting nodes»
  }
  «use nxt» // ← no accounting for nxt being one of the above deleted nodes

This was effectively a use-after-free, diagnosable by running under Address
Sanitizer. The solution in this change is to postpone all deletion until the
end.

Note that this does not fix the underlying sharp edge on which users can still
cut themselves: GVPR permits dangling references, allowing user scripts to
violate memory safety.

Gitlab: fixes #1702
Reported-by: Steve Roush
Suggested-by: Steve Roush