tests: abstract “using ASan?” into a function
The main intent of this change, apart from an improvement in readability at the
call site, is to make this check work for ASan-instrumented builds being tested
locally. The check no longer uses a CI-defined environment variable to figure
out whether ASan is in use, but a more agnostic binary inspection technique.