common point_init: fix incorrect construction of (-1, -1) points
Certain input graphs would fail an assertion:

  Assertion failed: (r->boundary[i] <= r->boundary[NUMDIMS + i]), function
    RTreeInsert, file index.c, line 192.

Rectangle construction was failing because it was based on points that claimed
to have width and height -72.0. This was the result of a bug that was introduced
in 8f2afe047367866179b7814a29aa4286e30d238c. What that change failed to take
into account was that the prior default of `MAXDOUBLE` ensured that the follow
up `w = fmin(w, h)`, in the presence of an unset width but set height, clamped
the width to the height’s value.

This change effectively reverts 8f2afe047367866179b7814a29aa4286e30d238c,
accounting for `MAXDOUBLE` having been replaced with the ISO C standard
`DBL_MAX` in the interim.

Gitlab: fixes #2613
Reported-by: cobizobi