cgraph _agstrcanon: fix missed escaping of double quote characters
When a double quote followed an escape sequence (e.g. “foo\n"bar”), the state
machine in this function would incorrectly conclude the quote had already been
escaped and did not need a “\” prepended. The problem stemmed from the escape
character itself (“n” in the example previously given) not causing the
`part_of_escape` flag to be reset during its own processing. This bug was
introduced in 3b81ca48fefe1bdcf30cbd3615ebf6dad60fb1db.

The state machine in this function seems unnecessarily complex for what it is
doing. And it is not obviously correct, as evidenced by both
3b81ca48fefe1bdcf30cbd3615ebf6dad60fb1db and this fix. It may be wise to rewrite
this into something simpler in future.

Gitlab: fixes #2614
Reported-by: Thomas Braun