Move doc/infosrc/attrs to cmd/gvedit/attrs.txt
The file is no longer used in doc/infosrc at all anymore, and the
gvedit build was copying it to be installed as 'attrs.txt' in the
gvedit datadir. Instead, have it live inside the gvedit source dir,
since that's the only place it's still used for anything.

Include attrs.txt in EXTRA_DIST

Remove attrs.txt from .gitignore