CI: fix: re-instate 'gvmap.sh' detection skipping on MinGW
Python 3.12 made some changes to how `shutil.which` functions that caused shell
scripts like gvmap.sh to now be detectable on Windows.
97b0c0b991dfad9c8b434576ea0bdd350143c315 explains some of the backstory to this.
18ebd129afcddc94cbc9ece6c60f2526bf29bf65 fixed various CI failures that began
occurring when this Python change propagated to MinGW in CI.

CI has now started once again failing on MinGW. Inspecting the diff between a
passing and failing run reveals Python has moved from 3.12.7-2 to 3.12.7-3. The
diff between these two appears to be the backporting of a claimed fix to
`shutil.which`.¹ Discussion of the underlying issue² seems to point to
significant disagreement between Python contributors on what the desirable
behavior of `shutil.which` is in these scenarios. It is also not clear to me how
gvmap.sh ends up non-executable (and thus affected by this) given its CMake
installation rule seems to add executability.

Rather than trying to continue accommodating Python changing its `shutil.which`
semantics, this change skips testing of this scenario.

¹ https://github.com/msys2/MINGW-packages/commit/32bd97ef8af58cdc36b9e796e35342d921d9b3a5
² https://github.com/python/cpython/issues/127001