fix a buffer overflow test case
This test was falsely failing because `dot("svg", …)` tries to UTF-8-decode
Graphviz output and triggers an exception because there are invalid UTF-8 bytes
in the output that have propagated from invalid UTF-8 bytes in the input. The
actual failure that was originally reported was a buffer overflow. Trying to
reproduce this in the present state triggers not a buffer overflow but a
use-after-free. This change tries to rephrase this test to capture all of this
and check for the actual intended failure scenarios.

Gitlab: #1472
Reported-by: Google Autofuzz project