add sandboxing wrapper
It seems common for users to process untrusted input in Graphviz. Some even host
Graphviz on internet-facing servers and let anonymous third parties submit
input. In a contemporary environment, this can be achieved safely with
mechanisms like virtualization and Web Assembly. If mechanisms like these are
unavailable to users, the script added in this commit provides a fallback
solution.

It should be straightforward to add support for sandboxing mechanisms beyond
Bubblewrap in future. For example, on macOS the logical mechanism would be
`sandbox-exec`.

In contrast to other Graphviz files, this new script has been designated as
public domain. There seems little value in retaining copyright over something
like this.

Gitlab: #2257