add a test case for failed loading of EXIF-prefixed JPEG
The JPEG in this commit was created by making a 1×1 white square in GIMP, adding
the metadata document title “1x1”, exporting this to JPEG with EXIF enabled,
then using a hex editor to swap the 0xE0 and 0xE1 segments¹ in the file. I have
not been able to find a JPEG-creating program on Linux that will create JPEGs
with the EXIF segment first (the root cause of this problem), something that
reportedly Apple iPhones do.

¹ https://en.wikipedia.org/wiki/JPEG#Syntax_and_structure

Gitlab: #2619
Reported-by: Ely Schoenfeld