relax conditions for recognizing a JPEG image
When trying to detect the format of an included image, Graphviz would look for
the magic bytes `\xff\xd8\xff\xe0` as indicative of a JPEG. This was overly
strict. While `\xff\xd8` must come first, JPEGs are made up of `\xff`-beginning
“segments”¹, one of which is `\xff\xe0` (“APP0”). All programs I have tried
produce JPEGs with an “APP0” segment in this position, yet it is technically
valid to have a different segment here, and reportedly there are systems in the
wild that write such JPEGs.

This change relaxes the JPEG check to “Start Of Image” followed by the `\xff`
byte beginning the next segment. This may be too loose, but users are not
expected to reference arbitrary non-image files within a graph, so this
recognition logic does not need to be precise.

¹ https://en.wikipedia.org/wiki/JPEG#Syntax_and_structure

Gitlab: #2619
Reported-by: Ely Schoenfeld