CI: remove use of Ccache on Cygwin
Of the three Cygwin CI jobs:

  1. windows-cygwin-build
  2. windows-cygwin-build-using-autogen
  3. windows-cygwin-cmake-build

both (1) and (3) were using Ccache incorrectly in a way that was effectively
useless. `CCACHE_DIR` was set to the relative path ccache-cache, but both these
jobs `cd`-ed somewhere else before beginning the build. As a result, Ccache
stored its results to a different path (graphviz-${GV_VERSION}/ccache-cache for
(1) and build/ccache-cache for (3)) than what Gitlab’s cache saving logic
expected. This could be seen in their logs:

  WARNING: ccache-cache: no matching files. Ensure that the artifact path is
    relative to the working directory (C:\GitLab-Runner\builds\smattr\graphviz)

In further runs, (1) and (3) would spend time restoring an empty directory.

This has been broken for a long time (forever?), so empirically Ccache is
inessential here. CI timing has some variance, but on a sample run of the state
before and after this change:

  ┌────────────────────────────────────┬────────┬────────┐
  │ job                                │ before │ after  │
  ├────────────────────────────────────┼────────┼────────┤
  │ windows-cygwin-build               │ 33m19s │ 30m30s │
  │ windows-cygwin-build-using-autogen │ 34m02s │ 33m18s │
  │ windows-cygwin-cmake-build         │ 18m11s │ 17m01s │
  └────────────────────────────────────┴────────┴────────┘