GD plugin gd_loadimage_cairo: fix 'free' of invalid pointer
The `data` pointer is incremented within this function. So passing it to `free`
was not passing the originally allocated address but rather some offset into the
allocated memory. This bug was introduced in
3677b2efb4dced4ec1befea640e4bc9860b2173b.

It is unclear to me why this was not caught in CI. The bug did not appear to
cause any crashes, even with input that certainly exercises this code path. The
bug was discovered by running various parts of the test suite locally with ASan
enabled.

Gitlab: #2619
Reported-by: Address Sanitizer
Reported-by: Ely Schoenfeld