add a test case for an xdot crash
This does not reproduce on any CI platform, even when rewinding to 2.44.1, the
Graphviz version it was reported against. However discussion on the Gitlab issue
suggests we were once able to reproduce it. So perhaps something has changed in
the platform environments since then. It seems wise to have a test case for this
in CI as we have an incomplete understanding of the problem.

Gitlab: #1896
Reported-by: nstander