CMake: add a plugin for outputing ASCII art
As discussed in the comments in this commit, rendering the AA-lib default way
did not seem to work. Excess lower and right white space was included in a way
that suggested AA-lib was miscalculating something internally. This is even
after accounting for it mapping 4 pixels onto 1 terminal character. Beyond
stripping this excess space, our work around has the beneficial side effect that
no ANSI terminal control characters are printed. The output can be piped to a
file, saving something reasonable looking.

Gitlab: closes #2421
Reported-by: Alexander Wessel