gml2gv: replace cdt queue with generic list for attribute lists
The generic list is more cache friendly and performant. However the main
motivation for this is moving towards removing `Dtqueue`.

We need to make `gmlattr.lp` a `void *` rather than a properly typed pointer
because there is a circular dependency involved. `gmlattr.lp` is an `attrs_t *`
but the definition of `attrs_t` needs the `gmlattr` definition to be seen first.
In future it might make sense to name the (generated) struct in
lib/cgraph/list.h so that it could be used in forward declarations. I.e. this
member could become something like `struct attrs *lp`.