gml2gv: replace cdt queue with generic list for graph lists
The generic list is more cache friendly and performant. However the main
motivation for this is moving towards removing `Dtqueue`.

We need to make `gmlgraph.graphlist` a `void *` rather than a properly typed
pointer because there is a circular dependency involved. `gmlgraph.graphlist` is
a `graphs_t *` but the definition of `graphs_t` needs the `gmlgraph` definition
to be seen first. In future it might make sense to name the (generated) struct
in lib/cgraph/list.h so that it could be used in forward declarations. I.e. this
member could become something like `struct graphs *graphlist`.