API BREAK: cgraph: remove unused 'agnew'
This seems to have been a part of the legacy libagraph that was carried over
into libcgraph without any uses. The last use of it was removed in
fb974a86dfae0917a8a5646864eb0e39151c758c. Debian Code Search also finds no uses
of this in the wild.¹

¹ https://codesearch.debian.net/search?q=%5Cbagnew%5Cb&literal=0

Gitlab: #1549