expr: remove 'Expr_t.id' which is always ""
The way this was described and used was both incorrect and inaccurate. The
`Expr_t.id` field was being set as a pointer to the end of its allocation. But
the actual `id` constructed earlier in `exccopen` was never copied into this
location. This was immaterial as both the `id` data and the existing sequence of
two zero bytes evaluate to the empty string. It is unclear to me why the
allocation was adding 2 and not 1 here. The documentation of “prefix + _” also
seems to have been incorrect; there is no “_” character involved.